pyproject.toml
[tool.poetry]
name = "ghostpost_backend"
version = "0.1.0"
description = ""
authors = ["Justin Rhuebottom"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
django-rest-framework = "^0.1.0"
pip = "^20.2.3"
install = "^1.3.4"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
